# Replace inspect.stack() with traceback.extract_stack() to solve performance bottleneck